wip: shift to using a shared library
