+++ app/src/main/java/com/chesire/malime/view/login/library/PrimeLibraryViewModel.kt
@@ -4,21 +4,21 @@ import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.LiveDataReactiveStreams
-import com.chesire.malime.kitsu.models.KitsuItem
-import com.chesire.malime.kitsu.repositories.KitsuLibrary
+import com.chesire.malime.core.models.MalimeModel
+import com.chesire.malime.core.repositories.Library
 import io.reactivex.BackpressureStrategy
 
 class PrimeLibraryViewModel(
     context: Application,
-    private val kitsuLibrary: KitsuLibrary
+    private val library: Library
 ) : AndroidViewModel(context) {
 
-    var myLibrary: LiveData<List<KitsuItem>> = LiveDataReactiveStreams.fromPublisher(
-        kitsuLibrary.getLibrary().toFlowable(BackpressureStrategy.ERROR)
+    var myLibrary: LiveData<List<MalimeModel>> = LiveDataReactiveStreams.fromPublisher(
+        library.observeLibrary().toFlowable(BackpressureStrategy.ERROR)
     )
 
     fun updateLibrary() {
         // will want to get updates on this, I.E %, number of items etc
-        kitsuLibrary.updateLibrary()
+        library.updateLibraryFromApi()
     }
 }
