+++ app/src/main/java/com/chesire/malime/view/login/library/PrimeLibraryViewModelFactory.kt
@@ -3,12 +3,12 @@ package com.chesire.malime.view.login.library
 import android.app.Application
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
-import com.chesire.malime.kitsu.repositories.KitsuLibrary
+import com.chesire.malime.core.repositories.Library
 import java.io.InvalidClassException
 
 class PrimeLibraryViewModelFactory(
     private val application: Application,
-    private val kitsuLibrary: KitsuLibrary
+    private val library: Library
 ) :
     ViewModelProvider.NewInstanceFactory() {
     @Suppress("UNCHECKED_CAST")
@@ -16,7 +16,7 @@ class PrimeLibraryViewModelFactory(
         when {
             modelClass.isAssignableFrom(PrimeLibraryViewModel::class.java) -> return PrimeLibraryViewModel(
                 application,
-                kitsuLibrary
+                library
             ) as T
             else -> throw InvalidClassException("")
         }
