+++ malime.core/src/main/java/com/chesire/malime/core/api/MalimeApi.kt
@@ -1,0 +1,4 @@
+package com.chesire.malime.core.api
+
+interface MalimeApi {
+}
